BUMCHEO-692 add api to swap, move chapter

+++ Mcsg.Comic.Api/Requests/Comics/ComicChapterOrderSwapR.cs
@@ -2,6 +2,6 @@
 
 public class ComicChapterOrderSwapR
 {
-    public int Order1 { get; set; }
-    public int Order2 { get; set; }
+    public float Order1 { get; set; }
+    public float Order2 { get; set; }
 }
